requirements.txt
torch>=1.13
torchaudio
openai
whisper
faster-whisper==1.0.2
modelscope
huggingface
huggingface_hub
funasr
onnxruntime
hyperpyyaml
Flask
gevent-websocket
flask_sockets
aiohttp-wsgi
uvicorn[standard]
soundfile
uvloop
aiofiles
webrtcvad
python-dotenv
edge-tts